+++ src/containers/Sidebar/index.js
@@ -1,5 +1,5 @@
 import React from 'react';
-
+import { connect } from 'react-redux';
 import {
 	AppSidebar,
 	AppSidebarFooter,
@@ -15,10 +15,21 @@ const Sidebar = (props) => {
 		<AppSidebar fixed display={props.display}>
 			<AppSidebarHeader />
 			<AppSidebarForm />
-			<SidebarNavItems navConfig={props.navigation.getItems().items} />
+			<SidebarNavItems
+				navConfig={props.navigation.getItems().items}
+				unauthorizedNavItems={props.unauthorizedNavItem}
+				unauthorizedNavChildren={props.unauthorizedNavChildren}
+			/>
 			<AppSidebarFooter />
 		</AppSidebar>
 	)
 }
 
-export default Sidebar;
+function mapStateToProps(state) {
+	return {
+		unauthorizedNavItem: state.auth?.unauthorizedNavItem,
+		unauthorizedNavChildren: state.auth?.unauthorizedNavChildren
+	};
+}
+
+export default connect(mapStateToProps)(Sidebar);