+++ CHANGELOG.md
@@ -52,6 +52,8 @@
 
 - Add documentation for Microservices (INDIGO Sprint 210820, [!159](https://github.com/TeskaLabs/asab-webui/pull/159))
 
+- Add option to hide sidebar children based on resource (INDIGO Sprint 210809, [!150](https://github.com/TeskaLabs/asab-webui/pull/150))
+
 
 ### Refactoring